# Generic modules over the integers and fields