--- a/src/Modules/UngradedModules/FreeMod.jl
+++ b/src/Modules/UngradedModules/FreeMod.jl
@@ -24,8 +24,8 @@ end
 
 
 @doc raw"""
-    free_module_sparse(R::ZZRing, n::Int, name::VarName = :e; cached::Bool = false)
-    free_module_sparse(R::Field, n::Int, name::VarName = :e; cached::Bool = false)
+    free_module(::Type{<:FreeMod}, R::Union{ZZRing, Field, MPolyRing, MPolyQuoRing, MPolyLocRing, MPolyQuoLocRing},
+                n::Int, name::VarName = :e; cached::Bool = false)
 
 Construct a free module of rank `n` over the ring `R` using a sparse implementation
 compatible with the generic modules framework, in cases where the `free_module` constructor 
@@ -35,7 +35,7 @@ The string `name` specifies how the basis vectors are printed.
 
 # Examples
 ```jldoctest
-julia> F = free_module_sparse(ZZ, 3, "f")
+julia> F = free_module(FreeMod, ZZ, 3, "f")
 Free module of rank 3 over integer ring
 
 julia> F[1]
@@ -43,14 +43,14 @@ f[1]
 
 julia> K = GF(7);
 
-julia> FK = free_module_sparse(K, 2)
+julia> FK = free_module(FreeMod, K, 2)
 Free module of rank 2 over K
 
 julia> FK[1]
 e[1]
 ```
 """
-free_module_sparse(R::Union{ZZRing, Field, MPolyRing, MPolyQuoRing, MPolyLocRing, MPolyQuoLocRing},
+free_module(::Type{<:FreeMod}, R::Union{ZZRing, Field, MPolyRing, MPolyQuoRing, MPolyLocRing, MPolyQuoLocRing},
     n::Int, name::VarName = :e; cached::Bool = false) = FreeMod(R, n, name, cached=cached)
 
 @doc raw"""