… docs/src/CommutativeAlgebra/ModulesOverMultivariateRings/free_modules.md
… docs/src/CommutativeAlgebra/ModulesOverMultivariateRings/subquotients.md
… docs/src/CommutativeAlgebra/homological_algebra.md
… src/Modules/UngradedModules/Presentation.jl
… src/Modules/UngradedModules/SubModuleOfFreeModule.jl
… src/exports.jl
… test/Modules/UngradedModules.jl