Make the generic modules work over the integers and fields. Also singles out the abstraction layer for the backend of the generic modules. The respective functions have atomic in their name. In a second iteration, the implementations of the atomic functions will be moved to separate files, and documentation of the abstraction layer will be added.