@HechtiDerLachs 2025-06-13
Timings for the local cohomology test suite: 

With this PR:
```julia
Test Summary:                  | Pass  Total     Time
first local cohomology example |    9      9  2m01.7s
Test Summary:      | Pass  Total     Time
hartshorne example |    6      6  2m39.6s
Test Summary:              | Pass  Total   Time
local cohomology of module |    1      1  58.1s
```
On master:
```julia
julia> include("experimental/InjectiveResolutions/test/local_cohomology.jl")
Test Summary:                  | Pass  Total     Time
first local cohomology example |    9      9  2m05.5s
Test Summary:      | Pass  Total     Time
hartshorne example |    6      6  2m46.1s
Test Summary:              | Pass  Total     Time
local cohomology of module |    1      1  1m02.1s
```